Ran `2to3 -w .` to add basic py3 support

+++ pypytools/testing/test_fakecython.py
@@ -6,8 +6,8 @@ class TestFakeCython:
     def test_fakecython(self):
         fake = FakeCython()
         @fake.cfunc
-        @fake.locals(a=long, b=long)
-        @fake.returns(long)
+        @fake.locals(a=int, b=int)
+        @fake.returns(int)
         def foo(a, b):
             return a+b
         #
